Move resources hub to homepage
